# Replace pandas str operation with list comprehension