Another speed improvement in the logfile parsing function.  Pandas's handling of string types is not really optimized compared with just normal python, so a list comprehension is actually faster than using the pandas vectorized functions.

Old approach: `df['linetype'] = df[0].str[0]`
```python
In [2]: %time _ = gemlog.ReadGem_v0_9_single('/home/kevin/Downloads/FILE0003.077')
CPU times: user 1.49 s, sys: 162 ms, total: 1.65 s
Wall time: 1.66 s
```

New approach: `df['linetype'] = [value[0] for value in df[0].str[0]`
```python
In [2]: %time _ = gemlog.ReadGem_v0_9_single('/home/kevin/Downloads/FILE0003.077')
CPU times: user 1.17 s, sys: 120 ms, total: 1.29 s
Wall time: 1.29 s
```